# server/schema/openapi.yaml
openapi: 3.0.3
info:
  title: server
  description: ''
  license:
    name: ''
  version: 0.1.0
paths:
  /api/getLogsByRange:
    post:
      tags:
      - Logs
      operationId: get_logs_by_range
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GetLogsByRangeModel'
        required: true
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetLogsByRangeResponse'
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/getLogsCount:
    get:
      tags:
      - Logs
      operationId: get_logs_count
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetLogsCountResponse'
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/insertLog:
    post:
      tags:
      - Logs
      operationId: insert_log
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogDto'
        required: true
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                default: null
                nullable: true
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
components:
  schemas:
    ApiError:
      type: object
      required:
      - message
      properties:
        message:
          type: string
    GetLogsByRangeModel:
      type: object
      required:
      - start
      - end
      properties:
        end:
          type: integer
          format: int64
        start:
          type: integer
          format: int64
    GetLogsByRangeResponse:
      type: object
      required:
      - logs
      properties:
        logs:
          type: array
          items:
            $ref: '#/components/schemas/Log'
    GetLogsCountResponse:
      type: object
      required:
      - count
      properties:
        count:
          type: integer
          format: int64
    Log:
      type: object
      required:
      - time
      - level
      - request_id
      - title
      properties:
        additional_data:
          nullable: true
        controller_name:
          type: string
          nullable: true
        level:
          $ref: '#/components/schemas/Loglevel'
        message:
          nullable: true
        request_id:
          type: string
          format: uuid
        service_name:
          type: string
          nullable: true
        time:
          type: string
          format: date-time
        title:
          type: string
    LogDto:
      type: object
      required:
      - level
      - title
      properties:
        additional_data:
          nullable: true
        controller_name:
          type: string
          nullable: true
        level:
          $ref: '#/components/schemas/Loglevel'
        message:
          nullable: true
        request_id:
          type: string
          format: uuid
          nullable: true
        service_name:
          type: string
          nullable: true
        title:
          type: string
    Loglevel:
      type: string
      enum:
      - debug
      - info
      - warning
      - error
      - critical